pyproject.toml
[project]
name = "airflow-dbt-dev"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow==2.11.0",
    "apache-airflow-providers-amazon==9.12.0",
    "astronomer-cosmos==1.10.2",
    "pandas==2.3.2",
    "pyiceberg[glue,s3fs,pyarrow]~=0.9.1",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "loguru==0.7.3",
]

[tool.setuptools.packages.find]
exclude = ["logs*", "config*", "plugins*", "__pycache__*"]
